Refine Shape ownership; refactor test pictures
Allow both owned and borrowed Shape parameters, following changes in
kurbo.

Reduce duplication of test picture code by putting it in a separate
crate. Over time, it is expected that crate will grow test and benchmark
functionality, in addition to just being a repository of pictures.

Partially addresses #17
