+++ piet-cairo/Cargo.toml
@@ -17,6 +17,9 @@ version = "0.5.0"
 # We don't need glib
 default-features = false
 
+[dev-dependencies]
+piet-test = { path = "../piet-test" }
+
 [dev-dependencies.cairo-rs]
 version = "0.5.0"
 features = ["png"]
