+++ piet-direct2d/Cargo.toml
@@ -19,3 +19,4 @@ directwrite = "0.1.4"
 direct3d11 = "0.1.7"
 dxgi = "0.1.7"
 image = "0.20.1"
+piet-test = { path = "../piet-test" }
