+++ piet-direct2d/examples/basic.rs
@@ -9,10 +9,10 @@ use direct3d11::flags::{BindFlags, CreateDeviceFlags};
 use direct3d11::helpers::ComWrapper;
 use dxgi::flags::Format;
 
-use kurbo::{Affine, BezPath, Line, Vec2};
-
-use piet::{FillRule, FontBuilder, RenderContext, TextLayout, TextLayoutBuilder};
+use piet::RenderContext;
 use piet_direct2d::D2DRenderContext;
+
+use piet_test::draw_test_picture;
 
 const TEXTURE_WIDTH: u32 = 400;
 const TEXTURE_HEIGHT: u32 = 200;
@@ -22,65 +22,10 @@ const TEXTURE_HEIGHT_S: usize = TEXTURE_HEIGHT as usize;
 
 const HIDPI: f32 = 2.0;
 
-// Note: this could be a Shape.
-fn star(center: Vec2, inner: f64, outer: f64, n: usize) -> BezPath {
-    let mut result = BezPath::new();
-    let d_th = std::f64::consts::PI / (n as f64);
-    for i in 0..n {
-        let outer_pt = center + outer * Vec2::from_angle(d_th * ((i * 2) as f64));
-        if i == 0 {
-            result.moveto(outer_pt);
-        } else {
-            result.lineto(outer_pt);
-        }
-        result.lineto(center + inner * Vec2::from_angle(d_th * ((i * 2 + 1) as f64)));
-    }
-    result.closepath();
-    result
-}
-
-fn draw_pretty_picture<R: RenderContext>(rc: &mut R) {
-    rc.clear(0xFF_FF_FF);
-    let brush = rc.solid_brush(0x00_00_80_FF);
-    rc.stroke(&Line::new((10.0, 10.0), (100.0, 50.0)), &brush, 1.0, None);
-
-    let mut path = BezPath::new();
-    path.moveto((50.0, 10.0));
-    path.quadto((60.0, 50.0), (100.0, 90.0));
-    let brush = rc.solid_brush(0x00_80_00_FF);
-    rc.stroke(&path, &brush, 1.0, None);
-
-    let mut path = BezPath::new();
-    path.moveto((10.0, 20.0));
-    path.curveto((10.0, 80.0), (100.0, 80.0), (100.0, 60.0));
-    let brush = rc.solid_brush(0x00_00_80_C0);
-    rc.fill(&path, &brush, FillRule::NonZero);
-
-    let font = rc.new_font_by_name("Segoe UI", 12.0).build();
-    let layout = rc.new_text_layout(&font, "Hello piet!").build();
-    let w: f64 = layout.width().into();
-    let brush = rc.solid_brush(0x80_00_00_C0);
-    rc.draw_text(&layout, (80.0, 10.0), &brush);
-
-    rc.stroke(
-        &Line::new((80.0, 12.0), (80.0 + w, 12.0)),
-        &brush,
-        1.0,
-        None,
-    );
-
-    rc.save();
-    rc.transform(Affine::rotate(0.1));
-    rc.draw_text(&layout, (80.0, 10.0), &brush);
-    rc.restore();
-
-    let clip_path = star(Vec2::new(90.0, 45.0), 10.0, 30.0, 24);
-    rc.clip(&clip_path, FillRule::NonZero);
-    let layout = rc.new_text_layout(&font, "Clipped text").build();
-    rc.draw_text(&layout, (80.0, 50.0), &brush);
-}
-
 fn main() {
+    // TODO: make selectable
+    let test_picture_number = 0;
+
     // Create the D2D factory
     let d2d = direct2d::factory::Factory::new().unwrap();
     let dwrite = directwrite::factory::Factory::new().unwrap();
@@ -115,7 +60,7 @@ fn main() {
     context.set_dpi(96.0 * HIDPI, 96.0 * HIDPI);
     context.begin_draw();
     let mut piet_context = D2DRenderContext::new(&d2d, &dwrite, &mut context);
-    draw_pretty_picture(&mut piet_context);
+    draw_test_picture(&mut piet_context, test_picture_number);
     piet_context.finish();
     context.end_draw().unwrap();
