+++ piet-test/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "piet-test"
+version = "0.1.0"
+authors = ["Raph Levien <raph.levien@gmail.com>"]
+description = "Utilities for testing the piet 2D graphics abstraction."
+license = "MIT/Apache-2.0"
+edition = "2018"
+keywords = ["graphics", "2d"]
+
+[dependencies]
+kurbo = "0.2.0"
+piet = { path = "../piet" }
